+++ packages/ai/src/generate-text/generate-text.ts
@@ -36,7 +36,7 @@ import {
 import { standardizePrompt } from '../prompt/standardize-prompt';
 import { wrapGatewayError } from '../prompt/wrap-gateway-error';
 import { createUnifiedTelemetry } from '../telemetry/create-unified-telemetry';
-import type { Telemetry } from '../telemetry/telemetry-integration';
+import type { Telemetry } from '../telemetry/telemetry';
 import { TelemetryOptions } from '../telemetry/telemetry-options';
 import {
   LanguageModel,
