+++ packages/ai/src/telemetry/index.ts
@@ -1,3 +1,3 @@
 export type { TelemetryOptions } from './telemetry-options';
-export type { Telemetry } from './telemetry-integration';
+export type { Telemetry } from './telemetry';
 export { registerTelemetry } from './telemetry-registry';
