+++ packages/ai/src/telemetry/telemetry-options.ts
@@ -1,4 +1,4 @@
-import type { Telemetry } from './telemetry-integration';
+import type { Telemetry } from './telemetry';
 
 /**
  * Telemetry configuration.
